feat: add proper upgrade to version 3 in Lido
